requirements.txt
groq
python-dotenv
flask